[BEAM-1487, BEAM-3016] Address termination correctness issues in BufferingStreamObserver & BeamFnLoggingClient
The issue with BeamFnLoggingClient is that we can't arriveAndDeregister during termination since
the onReadyHandler may also arrive at the same time which is why we swap to using forced termination.
Also, I added code that would guarantee that log messages produced by the thread which is shutting
down are guaranteed to make it (this was being caught occassionally by the testLogging test).

The BufferingStreamObserver was incorrectly shutting down since it may attempt to enqueue something
into a full queue with a reading thread that has already exitted for some reason so it would loop
forever attempting to insert the poison pill.